pyproject.toml
[tool.poetry]
name = "first-screening"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask_sqlalchemy = "^2.4.4"
mysqlclient = "^2.0.1"
Flask-MySQLdb = "^0.2.0"
sqlalchemy = "^1.3.19"
sqlalchemy-utils = "^0.36.8"
PyJWT = "^1.7.1"
factory_boy = "^3.0.1"
mysql-connector = "^2.2.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"
bandit = "^1.6.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
